requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
certifi==2025.1.31
click==8.1.8
distro==1.9.0
fastapi==0.115.11
googleapis-common-protos==1.69.0
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lz4==4.4.3
openai==1.65.2
packaging==24.2
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
protobuf==5.29.3
protoc-gen-openapiv2==0.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
